requirements.txt
pyyaml
openai
pandas
streamlit